Move propagate_terminate_agreement to protocol
